RAG-backend/pyproject.toml
[project]
name = "rag-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "cohere>=5.20.0",
    "fastapi>=0.125.0",
    "openai>=2.13.0",
    "openai-agents>=0.6.3",
    "python-dotenv>=1.2.1",
    "qdrant-client>=1.16.2",
    "requests>=2.32.5",
    "uvicorn>=0.38.0",
    "google-generativeai>=0.8.0",
]
